Removed content-type from requests and moved it to SetUp
